requirements.txt
airflow==1.7.1.3
alabaster==0.7.9
alembic==0.8.10
anaconda-clean==1.0
anaconda-client==1.5.1
anaconda-navigator==1.3.1
appdirs==1.4.3
appnope==0.1.0
appscript==1.0.1
APScheduler==3.2.0
argcomplete==1.0.0
astroid==1.4.7
astropy==1.2.1
attrs==17.2.0
Automat==0.6.0
avro==1.8.1
awscli==1.11.35
awsebcli==3.11.0
Babel==1.3
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.4.0.2
bayesian-optimization==0.3.0
beautifulsoup4==4.5.1
bitarray==0.8.1
blaze==0.10.1
blessed==1.14.2
blinker==1.4
bokeh==0.12.1
boto==2.33.0
boto3==1.4.3
botocore==1.4.92
Bottleneck==1.1.0
branca==0.2.0
cassandra-driver==3.6.0
cdecimal==2.3
cement==2.8.2
certifi==2017.7.27.1
cffi==1.7.0
chardet==3.0.4
chartkick==0.4.2
chest==0.2.3
click==6.6
cloudpickle==0.2.1
clyent==1.2.2
colorama==0.3.7
conda==4.3.23
conda-build==1.20.3
configobj==5.0.6
configparser==3.5.0
confluent-kafka==0.9.1.2
constantly==15.1.0
contextlib2==0.5.3
croniter==0.3.14
cryptography==1.7.2
cssselect==1.0.1
cycler==0.10.0
Cython==0.24.1
cytoolz==0.8.0
dask==0.11.0
datashape==0.5.2
db.py==0.5.3
decorator==4.1.2
demjson==2.2.4
dill==0.2.5
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.5
django-braces==1.11.0
django-csvimport==2.11
django-highcharts==0.1.7
django-markup-deprecated==0.0.3
django-toolbelt==0.0.1
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
dynd==0.7.3.dev1
enum34==1.1.6
et-xmlfile==1.0.1
fastavro==0.14.3
fastcache==1.0.2
feedgenerator==1.9
filechunkio==1.8
filelock==2.0.6
Flask==0.10.1
Flask-Admin==1.4.0
Flask-Cache==0.13.1
Flask-Cors==2.1.2
flask-heroku==0.1.9
Flask-Login==0.2.11
Flask-Split==0.3.0
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
folium==0.3.0
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.15.2
futures==3.0.5
gdata==2.0.18
genson==0.1.0
geopy==1.11.0
get==0.0.0
gevent==1.1.2
google-api-python-client==1.6.1
googlefinance==0.7
greenlet==0.4.10
grin==1.2.1
gunicorn==19.4.5
h5py==2.6.0
healthgraph-api==0.3.0
HeapDict==1.0.0
httplib2==0.9.2
hyperlink==17.2.1
hypothesis==3.17.0
idna==2.6
imagesize==0.7.1
incremental==17.5.0
ipaddress==1.0.16
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.2.0
ipywidgets==5.2.2
itsdangerous==0.24
jdcal==1.2
jedi==0.9.0
Jinja2==2.8
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.0.0
jupyter-core==4.3.0
kafka-python==1.3.3
Keras==1.1.1
lazy-object-proxy==1.2.1
llvmlite==0.13.0
locket==0.2.0
lockfile==0.12.2
logging==0.4.9.6
lxml==3.6.4
Mako==1.0.6
Markdown==2.6.8
MarkupSafe==0.23
matplotlib==1.5.3
mistune==0.7.3
mock==2.0.0
mpmath==0.19
mrjob==0.5.7
multipledispatch==0.4.8
multiprocessing==2.6.2.1
mysql==0.0.1
MySQL-python==1.2.5
nb-anacondacloud==1.2.0
nb-conda==2.0.0
nb-conda-kernels==2.0.0
nbconvert==4.2.0
nbformat==4.4.0
nbpresent==3.0.2
neo4j-driver==1.4.0
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.2.3
numba==0.28.1+0.gfe99fbc.dirty
numexpr==2.6.1
numpy==1.11.2
oauth2client==4.0.0
odo==0.5.0
openpyxl==2.3.2
packaging==16.8
pandas==0.18.1
parsel==1.2.0
partd==0.3.6
path.py==0.0.0
pathlib2==2.1.0
pathspec==0.5.0
patsy==0.4.1
pbr==1.10.0
pelican==3.6.3
pep8==1.7.0
perl==0.1
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.3.1
pkginfo==1.3.2
plotly==2.0.15
ply==3.9
post==0.0.0
prettytable==0.7.2
prompt-toolkit==1.0.3
protobuf==3.0.0
psclient==2.0
psutil==4.3.1
psycopg2==2.6.1
ptyprocess==0.5.1
public==0.0.0
py==1.4.31
py2neo==2.0.8
pyasn1==0.1.9
pyasn1-modules==0.0.9
PyAudio==0.2.7
pybars3==0.9.3
pybikes==1.0
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
PyDispatcher==2.0.5
pyflakes==1.3.0
Pygments==2.1.3
PyHive==0.4.0
pylint==1.5.4
PyMeta3==0.5.1
pymongo==3.1.1
Pympler==0.5
PyMySQL==0.7.10
pyOpenSSL==16.2.0
pyparsing==2.2.0
pysqlite==2.8.3
pytest==2.9.2
python-daemon==2.1.2
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.3.3
qtconsole==4.2.1
QtPy==1.1.2
query-string==0.0.0
queuelib==1.4.2
redis==2.10.5
request==0.0.0
requests==2.18.4
rope==0.9.4
rsa==3.4.2
ruamel-yaml===-VERSION
s3transfer==0.1.10
sasl==0.2.1
schedule==0.4.3
scikit-image==0.12.3
scikit-learn==0.18.1
scipy==0.18.1
Scrapy==1.4.0
seaborn==0.7.1
selenium==3.3.3
semantic-version==2.5.0
service-identity==17.0.0
setproctitle==1.1.10
setupfiles==0.0.0
Shapely==1.5.17.post1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sockjs-tornado==1.0.3
Sphinx==1.4.6
spyder==3.0.2
SQLAlchemy==1.0.12
static3==0.7.0
statsmodels==0.6.1
sympy==1.0
tables==3.2.3.1
tabulate==0.7.5
tensorflow==0.11.0rc2
termcolor==1.1.0
terminado==0.6
Theano==0.8.2
thrift==0.10.0
thrift-sasl==0.2.1
toolz==0.8.0
tornado==4.5.1
traitlets==4.3.2
Twisted==17.5.0
twitter==1.17.1
tzlocal==1.3
unicodecsv==0.14.1
Unidecode==0.4.19
uritemplate==3.0.0
urllib==1.21.1
urllib3==1.22
virtualenv==15.0.3
w3lib==1.17.0
wcwidth==0.1.7
websocket-client==0.44.0
Werkzeug==0.11.5
widgetsnbextension==1.2.6
wrapt==1.10.6
WTForms==2.1
xgboost==0.6
xlrd==1.0.0
XlsxWriter==0.9.3
xlwings==0.10.0
xlwt==1.1.2
zope.interface==4.4.2
